+++ tests/database/migrations/2018_11_25_000005_create_testbench_products_table.php
@@ -12,11 +12,8 @@ final class CreateTestbenchProductsTable extends Migration
             // Composite primary key
             $table->string('barcode');
             $table->string('uuid');
-
             $table->string('name');
-
-            $table->unsignedInteger('color_id');
-
+            $table->unsignedBigInteger('color_id');
             $table->timestamps();
         });
     }
